+++ core/src/formatter.rs
@@ -979,7 +979,8 @@ mod tests {
             .parser(DelphiLogicalLineParser {})
             .file_formatter(IndentBasedOnLineNumber {})
             .reconstructor(DelphiLogicalLinesReconstructor::new(
-                ReconstructionSettings::new("\n".to_owned(), " ".to_owned(), " ".to_owned()),
+                ReconstructionSettings::new("\n".to_owned(), " ".to_owned(), " ".to_owned())
+                    .unwrap(),
             ))
             .build();
         run_test(
@@ -1014,7 +1015,8 @@ mod tests {
             .file_formatter(IndentBasedOnLineNumber {})
             .file_formatter(IndentSecondLine3SpacesIfNoNewLine {})
             .reconstructor(DelphiLogicalLinesReconstructor::new(
-                ReconstructionSettings::new("\n".to_owned(), " ".to_owned(), " ".to_owned()),
+                ReconstructionSettings::new("\n".to_owned(), " ".to_owned(), " ".to_owned())
+                    .unwrap(),
             ))
             .build();
         run_test(
@@ -1067,7 +1069,8 @@ mod tests {
             .file_formatter(IndentSecondLine3SpacesIfNoNewLine {})
             .line_formatter(RetainSpacesLogcialLinesOnNewLines {})
             .reconstructor(DelphiLogicalLinesReconstructor::new(
-                ReconstructionSettings::new("\n".to_owned(), " ".to_owned(), " ".to_owned()),
+                ReconstructionSettings::new("\n".to_owned(), " ".to_owned(), " ".to_owned())
+                    .unwrap(),
             ))
             .build();
         run_test(